requirements.txt
boto3==1.26.133
schedule==1.2.0
python-dotenv==1.0.0
requests==2.30.0